adding guard clause to delete method in review_app_target.rb
container and task_definition cannot update, if task is not exist.
task is only created, when PRs were created.
if error occurs when PR was created, task is not created.

+++ app/models/review_app_target.rb
@@ -13,6 +13,8 @@ class ReviewAppTarget
   end
 
   def delete
+    return if task.nil?
+
     task.stop
     task.destroy
     self
